refactor(clips): features 구조로 도메인 레이어 이동
