+++ src/app/(app)/recent/page.tsx
@@ -1,4 +1,4 @@
-import { RecentClipsPage } from "@/domains/clips/ui/RecentClipsPage";
+import { RecentClipsPage } from "@/features/clip/ui/RecentClipsPage";
 
 export default function Recent() {
   return <RecentClipsPage />;
