+++ src/app/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginPage } from "@/domains/auth/ui/LoginPage";
+import { LoginPage } from "@/features/auth/ui/LoginPage";
 
 export default function Login() {
   return <LoginPage />;
