+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import { LandingPage } from "@/domains/landing/ui/LandingPage";
+import { LandingPage } from "@/features/landing/ui/LandingPage";
 
 export default function Home() {
   return <LandingPage />;
