+++ src/features/clip/model/clip.ts
