+++ src/features/clip/ui/ClipItem.tsx
@@ -8,7 +8,7 @@ import {
   HiOutlineStar,
   HiStar,
 } from "react-icons/hi";
-import { Clip } from "@/domains/clips/model/clip";
+import { Clip } from "@/features/clip/model/clip";
 
 interface ClipItemProps {
   clip: Clip;
