+++ src/features/clip/ui/DeleteAllButton.tsx
