+++ src/features/clip/ui/EmptyState.tsx
