+++ src/features/clip/ui/RecentClipsPage.tsx
@@ -1,37 +1,40 @@
 "use client";
 
-import { useFavoriteClipsPage } from "@/domains/clips/hooks/useFavoriteClipsPage";
-import { ClipList } from "@/domains/clips/ui/ClipList";
-import { EmptyState } from "@/domains/clips/ui/EmptyState";
-import { FilterBar } from "@/domains/clips/ui/FilterBar";
+import { useRecentClipsPage } from "@/features/clip/hooks/useRecentClipsPage";
+import { ClipList } from "@/features/clip/ui/ClipList";
+import { DeleteAllButton } from "@/features/clip/ui/DeleteAllButton";
+import { EmptyState } from "@/features/clip/ui/EmptyState";
+import { FilterBar } from "@/features/clip/ui/FilterBar";
 
-export function FavoriteClipsPage() {
+export function RecentClipsPage() {
   const {
     activeFilter,
+    clearAll,
     copyToast,
     filteredClips,
     handleCopy,
-    handleToggleFavorite,
+    hasClips,
+    searchQuery,
     setActiveFilter,
-  } = useFavoriteClipsPage();
+    setSearchQuery,
+  } = useRecentClipsPage();
 
   return (
     <div className="bg-background flex h-full flex-col">
       <FilterBar
         activeFilter={activeFilter}
         onFilterChange={setActiveFilter}
+        searchQuery={searchQuery}
+        onSearchChange={setSearchQuery}
         showStatus={false}
         countLabel={`${filteredClips.length} clips`}
       />
       {filteredClips.length ? (
-        <ClipList
-          clips={filteredClips}
-          onCopy={handleCopy}
-          onToggleFavorite={handleToggleFavorite}
-        />
+        <ClipList clips={filteredClips} onCopy={handleCopy} />
       ) : (
         <EmptyState />
       )}
+      <DeleteAllButton disabled={!hasClips} onClick={clearAll} />
 
       {copyToast ? (
         <div
